feat(pr): extract the files logic from `github.service.ts`
